// packages/api-client/src/generated/billing/billing.ts
export type v1BillingPlansListResponse200 = {
  data: PaginatedPlanList
  status: 200
}

export type v1BillingPlansListResponse400 = {
  data: Problem
  status: 400
}

export type v1BillingPlansListResponse401 = {
  data: Problem
  status: 401
}

export type v1BillingPlansListResponse403 = {
  data: Problem
  status: 403
}

export type v1BillingPlansListResponse404 = {
  data: Problem
  status: 404
}

export type v1BillingPlansListResponse409 = {
  data: Problem
  status: 409
}

export type v1BillingPlansListResponse422 = {
  data: Problem
  status: 422
}

export type v1BillingPlansListResponse500 = {
  data: Problem
  status: 500
}
    
export type v1BillingPlansListResponseSuccess = (v1BillingPlansListResponse200) & {
  headers: Headers;
};
export type v1BillingPlansListResponseError = (v1BillingPlansListResponse400 | v1BillingPlansListResponse401 | v1BillingPlansListResponse403 | v1BillingPlansListResponse404 | v1BillingPlansListResponse409 | v1BillingPlansListResponse422 | v1BillingPlansListResponse500) & {
  headers: Headers;
};

export type v1BillingPlansListResponse = (v1BillingPlansListResponseSuccess | v1BillingPlansListResponseError)
// ...
export type v1BillingPlansRetrieveResponse200 = {
  data: V1BillingPlansRetrieve200
  status: 200
}

export type v1BillingPlansRetrieveResponse400 = {
  data: Problem
  status: 400
}

export type v1BillingPlansRetrieveResponse401 = {
  data: Problem
  status: 401
}

export type v1BillingPlansRetrieveResponse403 = {
  data: Problem
  status: 403
}

export type v1BillingPlansRetrieveResponse404 = {
  data: Problem
  status: 404
}

export type v1BillingPlansRetrieveResponse409 = {
  data: Problem
  status: 409
}

export type v1BillingPlansRetrieveResponse422 = {
  data: Problem
  status: 422
}

export type v1BillingPlansRetrieveResponse500 = {
  data: Problem
  status: 500
}
    
export type v1BillingPlansRetrieveResponseSuccess = (v1BillingPlansRetrieveResponse200) & {
  headers: Headers;
};
export type v1BillingPlansRetrieveResponseError = (v1BillingPlansRetrieveResponse400 | v1BillingPlansRetrieveResponse401 | v1BillingPlansRetrieveResponse403 | v1BillingPlansRetrieveResponse404 | v1BillingPlansRetrieveResponse409 | v1BillingPlansRetrieveResponse422 | v1BillingPlansRetrieveResponse500) & {
  headers: Headers;
};

export type v1BillingPlansRetrieveResponse = (v1BillingPlansRetrieveResponseSuccess | v1BillingPlansRetrieveResponseError)

export const getV1BillingPlansRetrieveUrl = (id: string,) => {


  

  return `/api/v1/billing/plans/${id}/`
}

export const v1BillingPlansRetrieve = async (id: string, options?: RequestInit): Promise<v1BillingPlansRetrieveResponse> => {
  
  return customFetch<v1BillingPlansRetrieveResponse>(getV1BillingPlansRetrieveUrl(id),
  {      
    ...options,
    method: 'GET'
    
    
  }
);}


export type v1BillingSubscriptionsListResponse200 = {
  data: PaginatedSubscriptionList
  status: 200
}

export type v1BillingSubscriptionsListResponse400 = {
  data: Problem
  status: 400
}

export type v1BillingSubscriptionsListResponse401 = {
  data: Problem
  status: 401
}

export type v1BillingSubscriptionsListResponse403 = {
  data: Problem
  status: 403
}

export type v1BillingSubscriptionsListResponse404 = {
  data: Problem
  status: 404
}

export type v1BillingSubscriptionsListResponse409 = {
  data: Problem
  status: 409
}

export type v1BillingSubscriptionsListResponse422 = {
  data: Problem
  status: 422
}

export type v1BillingSubscriptionsListResponse500 = {
  data: Problem
  status: 500
}
    
export type v1BillingSubscriptionsListResponseSuccess = (v1BillingSubscriptionsListResponse200) & {
  headers: Headers;
};
export type v1BillingSubscriptionsListResponseError = (v1BillingSubscriptionsListResponse400 | v1BillingSubscriptionsListResponse401 | v1BillingSubscriptionsListResponse403 | v1BillingSubscriptionsListResponse404 | v1BillingSubscriptionsListResponse409 | v1BillingSubscriptionsListResponse422 | v1BillingSubscriptionsListResponse500) & {
  headers: Headers;
};

export type v1BillingSubscriptionsListResponse = (v1BillingSubscriptionsListResponseSuccess | v1BillingSubscriptionsListResponseError)
// ...
export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse200 = {
  data: V1BillingSubscriptionsCreateCheckoutSessionCreate200
  status: 200
}

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse400 = {
  data: Problem
  status: 400
}

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse401 = {
  data: Problem
  status: 401
}

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse403 = {
  data: Problem
  status: 403
}

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse404 = {
  data: Problem
  status: 404
}

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse409 = {
  data: Problem
  status: 409
}

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse422 = {
  data: Problem
  status: 422
}

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse500 = {
  data: Problem
  status: 500
}
    
export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponseSuccess = (v1BillingSubscriptionsCreateCheckoutSessionCreateResponse200) & {
  headers: Headers;
};
export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponseError = (v1BillingSubscriptionsCreateCheckoutSessionCreateResponse400 | v1BillingSubscriptionsCreateCheckoutSessionCreateResponse401 | v1BillingSubscriptionsCreateCheckoutSessionCreateResponse403 | v1BillingSubscriptionsCreateCheckoutSessionCreateResponse404 | v1BillingSubscriptionsCreateCheckoutSessionCreateResponse409 | v1BillingSubscriptionsCreateCheckoutSessionCreateResponse422 | v1BillingSubscriptionsCreateCheckoutSessionCreateResponse500) & {
  headers: Headers;
};

export type v1BillingSubscriptionsCreateCheckoutSessionCreateResponse = (v1BillingSubscriptionsCreateCheckoutSessionCreateResponseSuccess | v1BillingSubscriptionsCreateCheckoutSessionCreateResponseError)
// ...
export type v1BillingSubscriptionsCurrentRetrieveResponse200 = {
  data: V1BillingSubscriptionsCurrentRetrieve200
  status: 200
}

export type v1BillingSubscriptionsCurrentRetrieveResponse400 = {
  data: Problem
  status: 400
}

export type v1BillingSubscriptionsCurrentRetrieveResponse401 = {
  data: Problem
  status: 401
}

export type v1BillingSubscriptionsCurrentRetrieveResponse403 = {
  data: Problem
  status: 403
}

export type v1BillingSubscriptionsCurrentRetrieveResponse404 = {
  data: Problem
  status: 404
}

export type v1BillingSubscriptionsCurrentRetrieveResponse409 = {
  data: Problem
  status: 409
}

export type v1BillingSubscriptionsCurrentRetrieveResponse422 = {
  data: Problem
  status: 422
}

export type v1BillingSubscriptionsCurrentRetrieveResponse500 = {
  data: Problem
  status: 500
}
    
export type v1BillingSubscriptionsCurrentRetrieveResponseSuccess = (v1BillingSubscriptionsCurrentRetrieveResponse200) & {
  headers: Headers;
};
export type v1BillingSubscriptionsCurrentRetrieveResponseError = (v1BillingSubscriptionsCurrentRetrieveResponse400 | v1BillingSubscriptionsCurrentRetrieveResponse401 | v1BillingSubscriptionsCurrentRetrieveResponse403 | v1BillingSubscriptionsCurrentRetrieveResponse404 | v1BillingSubscriptionsCurrentRetrieveResponse409 | v1BillingSubscriptionsCurrentRetrieveResponse422 | v1BillingSubscriptionsCurrentRetrieveResponse500) & {
  headers: Headers;
};

export type v1BillingSubscriptionsCurrentRetrieveResponse = (v1BillingSubscriptionsCurrentRetrieveResponseSuccess | v1BillingSubscriptionsCurrentRetrieveResponseError)
// ...
export type v1BillingSubscriptionsOverviewRetrieveResponse200 = {
  data: V1BillingSubscriptionsOverviewRetrieve200
  status: 200
}

export type v1BillingSubscriptionsOverviewRetrieveResponse400 = {
  data: Problem
  status: 400
}

export type v1BillingSubscriptionsOverviewRetrieveResponse401 = {
  data: Problem
  status: 401
}

export type v1BillingSubscriptionsOverviewRetrieveResponse403 = {
  data: Problem
  status: 403
}

export type v1BillingSubscriptionsOverviewRetrieveResponse404 = {
  data: Problem
  status: 404
}

export type v1BillingSubscriptionsOverviewRetrieveResponse409 = {
  data: Problem
  status: 409
}

export type v1BillingSubscriptionsOverviewRetrieveResponse422 = {
  data: Problem
  status: 422
}

export type v1BillingSubscriptionsOverviewRetrieveResponse500 = {
  data: Problem
  status: 500
}
    
export type v1BillingSubscriptionsOverviewRetrieveResponseSuccess = (v1BillingSubscriptionsOverviewRetrieveResponse200) & {
  headers: Headers;
};
export type v1BillingSubscriptionsOverviewRetrieveResponseError = (v1BillingSubscriptionsOverviewRetrieveResponse400 | v1BillingSubscriptionsOverviewRetrieveResponse401 | v1BillingSubscriptionsOverviewRetrieveResponse403 | v1BillingSubscriptionsOverviewRetrieveResponse404 | v1BillingSubscriptionsOverviewRetrieveResponse409 | v1BillingSubscriptionsOverviewRetrieveResponse422 | v1BillingSubscriptionsOverviewRetrieveResponse500) & {
  headers: Headers;
};

export type v1BillingSubscriptionsOverviewRetrieveResponse = (v1BillingSubscriptionsOverviewRetrieveResponseSuccess | v1BillingSubscriptionsOverviewRetrieveResponseError)
// ...
export type v1BillingSubscriptionsRetrieveResponse200 = {
  data: V1BillingSubscriptionsRetrieve200
  status: 200
}

export type v1BillingSubscriptionsRetrieveResponse400 = {
  data: Problem
  status: 400
}

export type v1BillingSubscriptionsRetrieveResponse401 = {
  data: Problem
  status: 401
}

export type v1BillingSubscriptionsRetrieveResponse403 = {
  data: Problem
  status: 403
}

export type v1BillingSubscriptionsRetrieveResponse404 = {
  data: Problem
  status: 404
}

export type v1BillingSubscriptionsRetrieveResponse409 = {
  data: Problem
  status: 409
}

export type v1BillingSubscriptionsRetrieveResponse422 = {
  data: Problem
  status: 422
}

export type v1BillingSubscriptionsRetrieveResponse500 = {
  data: Problem
  status: 500
}
    
export type v1BillingSubscriptionsRetrieveResponseSuccess = (v1BillingSubscriptionsRetrieveResponse200) & {
  headers: Headers;
};
export type v1BillingSubscriptionsRetrieveResponseError = (v1BillingSubscriptionsRetrieveResponse400 | v1BillingSubscriptionsRetrieveResponse401 | v1BillingSubscriptionsRetrieveResponse403 | v1BillingSubscriptionsRetrieveResponse404 | v1BillingSubscriptionsRetrieveResponse409 | v1BillingSubscriptionsRetrieveResponse422 | v1BillingSubscriptionsRetrieveResponse500) & {
  headers: Headers;
};

export type v1BillingSubscriptionsRetrieveResponse = (v1BillingSubscriptionsRetrieveResponseSuccess | v1BillingSubscriptionsRetrieveResponseError)
// ...
export type v1BillingSubscriptionsManageSubscriptionCreateResponseSuccess = (v1BillingSubscriptionsManageSubscriptionCreateResponse200) & {
  headers: Headers;
};
export type v1BillingSubscriptionsManageSubscriptionCreateResponseError = (v1BillingSubscriptionsManageSubscriptionCreateResponse400 | v1BillingSubscriptionsManageSubscriptionCreateResponse401 | v1BillingSubscriptionsManageSubscriptionCreateResponse403 | v1BillingSubscriptionsManageSubscriptionCreateResponse404 | v1BillingSubscriptionsManageSubscriptionCreateResponse409 | v1BillingSubscriptionsManageSubscriptionCreateResponse422 | v1BillingSubscriptionsManageSubscriptionCreateResponse500) & {
  headers: Headers;
};

export type v1BillingSubscriptionsManageSubscriptionCreateResponse = (v1BillingSubscriptionsManageSubscriptionCreateResponseSuccess | v1BillingSubscriptionsManageSubscriptionCreateResponseError)
// ...
export type v1BillingInvoicesListResponseSuccess = (v1BillingInvoicesListResponse200) & {
  headers: Headers;
};
export type v1BillingInvoicesListResponseError = (v1BillingInvoicesListResponse400 | v1BillingInvoicesListResponse401 | v1BillingInvoicesListResponse403 | v1BillingInvoicesListResponse404 | v1BillingInvoicesListResponse409 | v1BillingInvoicesListResponse422 | v1BillingInvoicesListResponse500) & {
  headers: Headers;
};

export type v1BillingInvoicesListResponse = (v1BillingInvoicesListResponseSuccess | v1BillingInvoicesListResponseError)
// ...
export type v1BillingInvoicesRetrieveResponseSuccess = (v1BillingInvoicesRetrieveResponse200) & {
  headers: Headers;
};
export type v1BillingInvoicesRetrieveResponseError = (v1BillingInvoicesRetrieveResponse400 | v1BillingInvoicesRetrieveResponse401 | v1BillingInvoicesRetrieveResponse403 | v1BillingInvoicesRetrieveResponse404 | v1BillingInvoicesRetrieveResponse409 | v1BillingInvoicesRetrieveResponse422 | v1BillingInvoicesRetrieveResponse500) & {
  headers: Headers;
};

export type v1BillingInvoicesRetrieveResponse = (v1BillingInvoicesRetrieveResponseSuccess | v1BillingInvoicesRetrieveResponseError)