// packages/api-client/src/generated/api.schemas.ts
export type SubscriptionRequestStatus = typeof SubscriptionRequestStatus[keyof typeof SubscriptionRequestStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SubscriptionRequestStatus = {
  active: 'active',
  canceled: 'canceled',
  incomplete: 'incomplete',
  incomplete_expired: 'incomplete_expired',
  past_due: 'past_due',
  trialing: 'trialing',
  unpaid: 'unpaid',
} as const;

export interface SubscriptionRequest {
  organization: string;
  plan: string;
  /** * `active` - Active
* `canceled` - Canceled
* `incomplete` - Incomplete
* `incomplete_expired` - Incomplete Expired
* `past_due` - Past Due
* `trialing` - Trialing
* `unpaid` - Unpaid */
  status?: SubscriptionRequestStatus;
  cancelAtPeriodEnd?: boolean;
  /** @nullable */
  gracePeriodEnd?: string | null;
  metadata?: unknown;
}
// ...
export type V1BillingPlansListParams = {
/**
 * A page number within the paginated result set.
 */
page?: number;
/**
 * Number of results to return per page.
 */
page_size?: number;
};

export type V1BillingPlansRetrieve200AllOf = {
  data?: Plan;
};

export type V1BillingPlansRetrieve200 = SuccessEnvelope & V1BillingPlansRetrieve200AllOf;

export type V1BillingSubscriptionsListParams = {
/**
 * A page number within the paginated result set.
 */
page?: number;
/**
 * Number of results to return per page.
 */
page_size?: number;
};
// ...
export type V1BillingInvoicesListParams = {
/**
 * A page number within the paginated result set.
 */
page?: number;
/**
 * Number of results to return per page.
 */
page_size?: number;
};